requirements.txt
-e git+https://github.com/garrettr/python-gnupg.git@search-keys#egg=gnupg
psutil==4.2.0
PyYAML==3.11
